requirements.txt
google-generativeai>=0.3.0
pydantic>=2.0.0
